+++ garage/envs/util.py
@@ -3,11 +3,11 @@ import numpy as np
 import theano
 
 from garage.envs.base import EnvSpec
-from garage.misc import ext
 from garage.misc import special
 from garage.spaces import Box as GarageBox
 from garage.spaces import Discrete as GarageDiscrete
 from garage.spaces import Product as GarageProduct
+from garage.theano.misc import tensor_utils
 
 __all__ = [
     'bounds', 'default_value', 'flat_dim', 'flatten', 'flatten_n', 'horizon',
@@ -77,17 +77,17 @@ def horizon(env):
 
 def new_tensor_variable(space, name, extra_dims):
     if isinstance(space, gym.spaces.Box):
-        return ext.new_tensor(
+        return tensor_utils.new_tensor(
             name=name, ndim=extra_dims + 1, dtype=theano.config.floatX)
     elif isinstance(space, gym.spaces.Discrete):
         if space.n <= 2**8:
-            return ext.new_tensor(
+            return tensor_utils.new_tensor(
                 name=name, ndim=extra_dims + 1, dtype='uint8')
         elif space.n <= 2**16:
-            return ext.new_tensor(
+            return tensor_utils.new_tensor(
                 name=name, ndim=extra_dims + 1, dtype='uint16')
         else:
-            return ext.new_tensor(
+            return tensor_utils.new_tensor(
                 name=name, ndim=extra_dims + 1, dtype='uint32')
     elif isinstance(space, gym.spaces.Tuple):
         dtypes = [
@@ -97,7 +97,7 @@ def new_tensor_variable(space, name, extra_dims):
         if dtypes and hasattr(dtypes[0], "as_numpy_dtype"):
             dtypes = [d.as_numpy_dtype for d in dtypes]
         common_dtype = np.core.numerictypes.find_common_type([], dtypes)
-        return ext.new_tensor(
+        return tensor_utils.new_tensor(
             name=name,
             ndim=extra_dims + 1,
             dtype=common_dtype,
