+++ garage/policies/gaussian_mlp_policy.py
@@ -5,7 +5,6 @@ import numpy as np
 import theano.tensor as TT
 
 from garage.core import Serializable
-from garage.misc import ext
 from garage.misc import logger
 from garage.misc.overrides import overrides
 from garage.policies import StochasticPolicy
@@ -14,6 +13,7 @@ from garage.theano.core import LasagnePowered
 from garage.theano.core import MLP
 from garage.theano.core import ParamLayer
 from garage.theano.distributions import DiagonalGaussian
+from garage.theano.misc import tensor_utils
 
 
 class GaussianMLPPolicy(StochasticPolicy, LasagnePowered):
@@ -111,7 +111,7 @@ class GaussianMLPPolicy(StochasticPolicy, LasagnePowered):
         LasagnePowered.__init__(self, [l_mean, l_log_std])
         super(GaussianMLPPolicy, self).__init__(env_spec)
 
-        self._f_dist = ext.compile_function(
+        self._f_dist = tensor_utils.compile_function(
             inputs=[obs_var],
             outputs=[mean_var, log_std_var],
         )
