+++ garage/regressors/gaussian_mlp_regressor.py
@@ -7,7 +7,6 @@ import theano.tensor as TT
 
 from garage.core import Serializable
 from garage.misc import logger
-from garage.misc.ext import compile_function
 from garage.misc.ext import iterate_minibatches_generic
 from garage.optimizers import LbfgsOptimizer
 from garage.optimizers import PenaltyLbfgsOptimizer
@@ -15,6 +14,7 @@ from garage.theano.core import LasagnePowered
 from garage.theano.core import MLP
 from garage.theano.core import ParamLayer
 from garage.theano.distributions import DiagonalGaussian
+from garage.theano.misc.tensor_utils import compile_function
 
 
 class GaussianMLPRegressor(LasagnePowered):
