+++ garage/theano/spaces/product.py
@@ -1,6 +1,6 @@
 """Spaces.Product for Theano."""
-from garage.misc import ext
 from garage.spaces import Product as GarageProduct
+from garage.theano.misc import tensor_utils
 
 
 class Product(GarageProduct):
@@ -14,7 +14,7 @@ class Product(GarageProduct):
         :param extra_dims: extra dimensions to be prepended
         :return: the created tensor variable
         """
-        return ext.new_tensor(
+        return tensor_utils.new_tensor(
             name=name,
             ndim=extra_dims + 1,
             dtype=self._common_dtype,
